server/pyproject.toml
[project]
name = "save_my_snippet"
version = "1.0"
description = "SaveMySnippet can be used to save all your code snippets without the worry of losing them when in need."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.29.0",
    "drf-spectacular-sidecar>=2025.12.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.0",
    "uvicorn-worker>=0.4.0",
    "uvicorn[standard]>=0.38.0",
]
